chore(api.js): add `@since` tag where appropriate
