+++ tooling/api/package.json
@@ -67,6 +67,7 @@
     "tslib": "2.4.0",
     "typedoc": "0.23.14",
     "typedoc-plugin-markdown": "3.13.6",
+    "typedoc-plugin-mdn-links": "^2.0.0",
     "typescript": "4.8.2"
   }
 }
