+++ tooling/api/src/app.ts
@@ -19,7 +19,7 @@ import { invokeTauriCommand } from './helpers/tauri'
  * const appVersion = await getVersion();
  * ```
  *
- * @returns A promise resolving to the application version.
+ * @since 1.0.0
  */
 async function getVersion(): Promise<string> {
   return invokeTauriCommand<string>({
@@ -38,7 +38,7 @@ async function getVersion(): Promise<string> {
  * const appName = await getName();
  * ```
  *
- * @returns A promise resolving to application name.
+ * @since 1.0.0
  */
 async function getName(): Promise<string> {
   return invokeTauriCommand<string>({
@@ -58,7 +58,7 @@ async function getName(): Promise<string> {
  * const tauriVersion = await getTauriVersion();
  * ```
  *
- * @returns A promise resolving to Tauri version.
+ * @since 1.0.0
  */
 async function getTauriVersion(): Promise<string> {
   return invokeTauriCommand<string>({
