+++ tooling/api/src/cli.ts
@@ -11,6 +11,9 @@
 
 import { invokeTauriCommand } from './helpers/tauri'
 
+/**
+ * @since 1.0.0
+ */
 interface ArgMatch {
   /**
    * string if takes value
@@ -24,11 +27,17 @@ interface ArgMatch {
   occurrences: number
 }
 
+/**
+ * @since 1.0.0
+ */
 interface SubcommandMatch {
   name: string
   matches: CliMatches
 }
 
+/**
+ * @since 1.0.0
+ */
 interface CliMatches {
   args: { [name: string]: ArgMatch }
   subcommand: SubcommandMatch | null
@@ -52,7 +61,7 @@ interface CliMatches {
  * }
  * ```
  *
- * @returns A promise resolving to the parsed arguments.
+ * @since 1.0.0
  */
 async function getMatches(): Promise<CliMatches> {
   return invokeTauriCommand<CliMatches>({
