+++ tooling/api/src/event.ts
@@ -11,11 +11,36 @@
 
 import * as eventApi from './helpers/event'
 import type {
-  EventName,
   EventCallback,
   UnlistenFn,
   Event
 } from './helpers/event'
+
+export type EventName = TauriEvent | string
+
+/**
+ * @since 1.1.0
+ */
+export enum TauriEvent {
+  WINDOW_RESIZED = 'tauri://resize',
+  WINDOW_MOVED = 'tauri://move',
+  WINDOW_CLOSE_REQUESTED = 'tauri://close-requested',
+  WINDOW_CREATED = 'tauri://window-created',
+  WINDOW_DESTROYED = 'tauri://destroyed',
+  WINDOW_FOCUS = 'tauri://focus',
+  WINDOW_BLUR = 'tauri://blur',
+  WINDOW_SCALE_FACTOR_CHANGED = 'tauri://scale-change',
+  WINDOW_THEME_CHANGED = 'tauri://theme-changed',
+  WINDOW_FILE_DROP = 'tauri://file-drop',
+  WINDOW_FILE_DROP_HOVER = 'tauri://file-drop-hover',
+  WINDOW_FILE_DROP_CANCELLED = 'tauri://file-drop-cancelled',
+  MENU = 'tauri://menu',
+  CHECK_UPDATE = 'tauri://update',
+  UPDATE_AVAILABLE = 'tauri://update-available',
+  INSTALL_UPDATE = 'tauri://update-install',
+  STATUS_UPDATE = 'tauri://update-status',
+  DOWNLOAD_PROGRESS = 'tauri://update-download-progress'
+}
 
 /**
  * Listen to an event from the backend.
@@ -33,8 +58,10 @@ import type {
  *
  * @param event Event name. Must include only alphanumeric characters, `-`, `/`, `:` and `_`.
  * @param handler Event handler callback.
- * @return A promise resolving to a function to unlisten to the event.
+ * @returns A promise resolving to a function to unlisten to the event.
  * Note that removing the listener is required if your listener goes out of scope e.g. the component is unmounted.
+ *
+ * @since 1.0.0
  */
 async function listen<T>(
   event: EventName,
@@ -62,9 +89,10 @@ async function listen<T>(
  * ```
  *
  * @param event Event name. Must include only alphanumeric characters, `-`, `/`, `:` and `_`.
- * @param handler Event handler callback.
  * @returns A promise resolving to a function to unlisten to the event.
  * Note that removing the listener is required if your listener goes out of scope e.g. the component is unmounted.
+ *
+ * @since 1.0.0
  */
 async function once<T>(
   event: EventName,
@@ -82,13 +110,13 @@ async function once<T>(
  * ```
  *
  * @param event Event name. Must include only alphanumeric characters, `-`, `/`, `:` and `_`.
- * @param [payload] Event payload
- * @returns
+ *
+ * @since 1.0.0
  */
 async function emit(event: string, payload?: unknown): Promise<void> {
   return eventApi.emit(event, undefined, payload)
 }
 
-export type { Event, EventName, EventCallback, UnlistenFn }
+export type { Event, EventCallback, UnlistenFn }
 
 export { listen, once, emit }
