+++ tooling/api/src/globalShortcut.ts
@@ -40,7 +40,8 @@ export type ShortcutHandler = (shortcut: string) => void
  *
  * @param shortcut Shortcut definition, modifiers and key separated by "+" e.g. CmdOrControl+Q
  * @param handler Shortcut handler callback - takes the triggered shortcut as argument
- * @returns
+ *
+ * @since 1.0.0
  */
 async function register(
   shortcut: string,
@@ -68,7 +69,8 @@ async function register(
  *
  * @param shortcuts Array of shortcut definitions, modifiers and key separated by "+" e.g. CmdOrControl+Q
  * @param handler Shortcut handler callback - takes the triggered shortcut as argument
- * @returns
+ *
+ * @since 1.0.0
  */
 async function registerAll(
   shortcuts: string[],
@@ -93,7 +95,8 @@ async function registerAll(
  * ```
  *
  * @param shortcut Array of shortcut definitions, modifiers and key separated by "+" e.g. CmdOrControl+Q
- * @returns A promise resolving to the state.
+ *
+ * @since 1.0.0
  */
 async function isRegistered(shortcut: string): Promise<boolean> {
   return invokeTauriCommand({
@@ -114,7 +117,8 @@ async function isRegistered(shortcut: string): Promise<boolean> {
  * ```
  *
  * @param shortcut shortcut definition, modifiers and key separated by "+" e.g. CmdOrControl+Q
- * @returns
+ *
+ * @since 1.0.0
  */
 async function unregister(shortcut: string): Promise<void> {
   return invokeTauriCommand({
@@ -134,7 +138,7 @@ async function unregister(shortcut: string): Promise<void> {
  * await unregisterAll();
  * ```
  *
- * @returns
+ * @since 1.0.0
  */
 async function unregisterAll(): Promise<void> {
   return invokeTauriCommand({
