+++ tooling/api/src/os.ts
@@ -57,6 +57,8 @@ type Arch =
  * The operating system-specific end-of-line marker.
  * - `\n` on POSIX
  * - `\r\n` on Windows
+ *
+ * @since 1.0.0
  * */
 const EOL = isWindows() ? '\r\n' : '\n'
 
@@ -68,6 +70,9 @@ const EOL = isWindows() ? '\r\n' : '\n'
  * import { platform } from '@tauri-apps/api/os';
  * const platformName = await platform();
  * ```
+ *
+ * @since 1.0.0
+ *
  */
 async function platform(): Promise<Platform> {
   return invokeTauriCommand<Platform>({
@@ -85,6 +90,8 @@ async function platform(): Promise<Platform> {
  * import { version } from '@tauri-apps/api/os';
  * const osVersion = await version();
  * ```
+ *
+ * @since 1.0.0
  */
 async function version(): Promise<string> {
   return invokeTauriCommand<string>({
@@ -102,6 +109,8 @@ async function version(): Promise<string> {
  * import { type } from '@tauri-apps/api/os';
  * const osType = await type();
  * ```
+ *
+ * @since 1.0.0
  */
 async function type(): Promise<OsType> {
   return invokeTauriCommand<OsType>({
@@ -120,6 +129,8 @@ async function type(): Promise<OsType> {
  * import { arch } from '@tauri-apps/api/os';
  * const archName = await arch();
  * ```
+ *
+ * @since 1.0.0
  */
 async function arch(): Promise<Arch> {
   return invokeTauriCommand<Arch>({
@@ -137,6 +148,8 @@ async function arch(): Promise<Arch> {
  * import { tempdir } from '@tauri-apps/api/os';
  * const tempdirPath = await tempdir();
  * ```
+ *
+ * @since 1.0.0
  */
 async function tempdir(): Promise<string> {
   return invokeTauriCommand<string>({
