+++ tooling/api/src/tauri.ts
@@ -30,6 +30,8 @@ function uid(): number {
  * The backend uses the identifier to `eval()` the callback.
  *
  * @return A unique identifier associated with the callback function.
+ *
+ * @since 1.0.0
  */
 function transformCallback(
   callback?: (response: any) => void,
@@ -53,7 +55,11 @@ function transformCallback(
   return identifier
 }
 
-/** Command arguments. */
+/**
+ * Command arguments.
+ *
+ * @since 1.0.0
+*/
 interface InvokeArgs {
   [key: string]: unknown
 }
@@ -69,6 +75,8 @@ interface InvokeArgs {
  * @param cmd The command name.
  * @param args The optional arguments to pass to the command.
  * @return A promise resolving or rejecting to the backend response.
+ *
+ * @since 1.0.0
  */
 async function invoke<T>(cmd: string, args: InvokeArgs = {}): Promise<T> {
   return new Promise((resolve, reject) => {
@@ -117,6 +125,8 @@ async function invoke<T>(cmd: string, args: InvokeArgs = {}): Promise<T> {
  * ```
  *
  * @return the URL that can be used as source on the webview.
+ *
+ * @since 1.0.0
  */
 function convertFileSrc(filePath: string, protocol = 'asset'): string {
   const path = encodeURIComponent(filePath)
